requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0
anthropic>=0.18.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pdfplumber>=0.10.0
PyPDF2>=3.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
python-dotenv>=1.0.0
